[feat] Add composition pattern to table component
