+++ data.json
@@ -1,21 +1,35 @@
 {
   "tableData": [
     {
-      "imageSrc": "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+      "imageSrc": "/image.jpg",
       "carName": "Mini Cooper 2020",
       "nextReservation": "11/07",
       "status": "Available",
       "rating": 5
     },
       {
-      "imageSrc": "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+      "imageSrc": "/image.jpg",
       "carName": "Mini Cooper 2020",
       "nextReservation": "11/07",
       "status": "Available",
       "rating": 5
       },
       {
-      "imageSrc": "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+        "imageSrc": "/image.jpg",
+      "carName": "Mini Cooper 2020",
+      "nextReservation": "11/07",
+      "status": "Available",
+      "rating": 6
+      },
+      {
+        "imageSrc": "/image.jpg",
+      "carName": "Mini Cooper 2020",
+      "nextReservation": "11/07",
+      "status": "Available",
+      "rating": 6
+      },
+      {
+        "imageSrc": "/image.jpg",
       "carName": "Mini Cooper 2020",
       "nextReservation": "11/07",
       "status": "Available",
