+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import Table from "../components/Table/index";
+import Table from "../components/Table/Table";
 
 const Home: React.FC = () => {
   const [isImageClicked, setIsImageClicked] = useState(false);
